+++ module/data/item/knack-maneuver.mjs
@@ -1,3 +1,4 @@
+import ManeuverData from "./maneuver.mjs";
 import ItemDescriptionTemplate from "./templates/item-description.mjs";
 import KnackTemplate from "./templates/knack-item.mjs";
 
@@ -5,7 +6,8 @@ import KnackTemplate from "./templates/knack-item.mjs";
  * Data model template with information on items that are used to represent custom active effects.
  */
 export default class KnackManeuverData extends KnackTemplate.mixin(
-    ItemDescriptionTemplate
+    ItemDescriptionTemplate,
+    ManeuverData
 ) {
 
     /** @inheritDoc */
