--- module/data/item/matrizen.mjs
@@ -1,27 +1,0 @@
-import SystemDataModel from "../abstract.mjs";
-import ItemDescriptionTemplate from "./templates/item-description.mjs";
-
-/**
- * Data model template with information on matrix items.
- */
-export default class MatrixData extends SystemDataModel.mixin(
-    ItemDescriptionTemplate
-) {
-
-    /** @inheritDoc */
-    static defineSchema() {
-        return this.mergeSchema( super.defineSchema(), {
-            
-        } );
-    }
-
-    /* -------------------------------------------- */
-    /*  Migrations                                  */
-    /* -------------------------------------------- */
-
-    /** @inheritDoc */
-    static migrateData( source ) {
-        super.migrateData( source );
-        // specific migration functions
-    }
-}
