+++ module/data/item/templates/ability.mjs
@@ -1,26 +1,22 @@
-import SystemDataModel from "../../abstract.mjs";
 import ClassTemplate from './class.mjs';
 import { DocumentUUIDField } from "../../fields.mjs";
+import TargetTemplate from "./targeting.mjs";
+import ActionTemplate from "./action.mjs";
 
 /**
  * Data model template with information on Ability items.
- * @property {string} action action type
  * @property {string} attribute attribute
- * @property {string} tier talent tier
- * @property {number} strain strain
+ * @property {object} source Class Source
+ * @property {string} source.class class
+ * @property {string} source.tier talent tier
  * @property {number} level rank
  */
-export default class AbilityTemplate extends SystemDataModel {
+export default class AbilityTemplate extends ActionTemplate.mixin( 
+    TargetTemplate 
+) {
     /** @inheritDoc */
     static defineSchema() {
         return this.mergeSchema( super.defineSchema(), {
-            action: new foundry.data.fields.StringField( {
-                required: true,
-                nullable: false,
-                blank: false,
-                initial: "standard",
-                label: "ED.Item.Ability.action"
-            } ),
             attribute: new foundry.data.fields.StringField( {
                 required: false,
                 nullable: true,
@@ -49,36 +45,6 @@ export default class AbilityTemplate extends SystemDataModel {
                 initial: 0,
                 integer: true,
                 label: "ED.Item.Ability.rank"
-            } ),
-            strain: new foundry.data.fields.NumberField( {
-                required: true,
-                nullable: false,
-                min: 0,
-                initial: 0,
-                integer: true,
-                label: "ED.Item.Ability.strain"
-            } ),
-            difficulty: new foundry.data.fields.SchemaField( {
-                target: new foundry.data.fields.StringField( {
-                    nullable: false,
-                    blank: false,
-                    initial: "none",
-                    label: "X.TargetDifficulty"
-                } ),
-                group: new foundry.data.fields.StringField( {
-                    nullable: false,
-                    blank: false,
-                    initial: "none",
-                    label: "X.GroupDifficulty"
-                } ),
-                fixed: new foundry.data.fields.NumberField( {
-                    required: true,
-                    nullable: false,
-                    min: 0,
-                    initial: 0,
-                    integer: true,
-                    label: "X.FixedDifficulty"
-                } ),
             } ),
         } );
     }
